+++ src/main/java/persistence/entity/loader/EntityLoader.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.loader;
 
 import jdbc.JdbcTemplate;
 import persistence.entity.exception.NotUniqueDataException;
