+++ src/test/java/persistence/entity/entityloader/EntityLoaderTest.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.entityloader;
 
 import database.DatabaseServer;
 import database.H2;
@@ -7,14 +7,17 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import persistence.entity.notcolumn.Person;
+import persistence.entity.loader.EntityLoader;
+import persistence.entity.manager.EntityManager;
+import persistence.entity.manager.EntityManagerImpl;
+import persistence.entity.testfixture.notcolumn.Person;
 import persistence.sql.ddl.CreateQueryBuilder;
 
 import static persistence.sql.ddl.common.TestSqlConstant.DROP_TABLE_USERS;
 
 class EntityLoaderTest {
 
-    private static final Logger logger = LoggerFactory.getLogger(EntityManagerTest.class);
+    private static final Logger logger = LoggerFactory.getLogger(EntityLoaderTest.class);
     private static DatabaseServer server;
     private static JdbcTemplate jdbcTemplate;
