+++ src/test/java/persistence/entity/entitypersister/EntityPersisterTest.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.entitypersister;
 
 import database.DatabaseServer;
 import database.H2;
@@ -7,7 +7,10 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import persistence.entity.notcolumn.Person;
+import persistence.entity.manager.EntityManager;
+import persistence.entity.manager.EntityManagerImpl;
+import persistence.entity.persister.EntityPersister;
+import persistence.entity.testfixture.notcolumn.Person;
 import persistence.sql.ddl.CreateQueryBuilder;
 
 import static persistence.sql.ddl.common.TestSqlConstant.DROP_TABLE_USERS;
