+++ src/test/java/persistence/entity/testfixture/basic/Dog.java
@@ -1,4 +1,4 @@
-package persistence.entity.basic;
+package persistence.entity.testfixture.basic;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
